Adds integration test for alternate pom and some refactoring.
